+++ lib/summary.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Summary
+end
+
+require "lib/sumary/parser"
