+++ lib/summary/archive.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Summary
+  class Archive
+    def initialize
+      @values = []
+    end
+
+    def add(pilot)
+      @values << record
+    end
+  end
+end
